He doesn't really waste demands behind him.
What to do? What should I do: Four months and I did not approve any withdrawal request? I face difficulty communicating and the people who hacked me are hiding the files in the phone. I can see them, but they are shaded, and when I open them, they are empty and white. After many attempts, I was able to monitor the accounts and identifiers of these fraudsters, as well as the sites that were hacking my accounts, from them, they are a professional organization, and I am alone and I have my phone only. I found some gaps, but it's useless that they do this from North America and they are in a military area? I don't know who I submit these recordings or who I will trust after what happened and what happens to me is a violation of my right to freedom of expression and publish content.